requirements.txt
torch==1.13.1
torchvision==0.14.1
seaborn==0.12.0
tensorboard
tqdm
wandb
hydra-core
matplotlib
